[CHANGE] move specific sources, repos, usecases from common to features
