+++ lib/src/common/data/repositories/authentication_repository_impl.dart
@@ -3,8 +3,8 @@ import 'dart:async';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:log/log.dart';
 import 'package:pocketbase/pocketbase.dart';
-import 'package:vocabualize/src/common/data/data_sources/authentication_data_source.dart';
-import 'package:vocabualize/src/common/data/data_sources/remote_connection_client.dart';
+import 'package:vocabualize/src/common/data/sources/authentication_data_source.dart';
+import 'package:vocabualize/src/common/data/sources/remote_connection_client.dart';
 import 'package:vocabualize/src/common/data/mappers/auth_mappers.dart';
 import 'package:vocabualize/src/common/domain/entities/app_user.dart';
 import 'package:vocabualize/src/common/domain/repositories/authentication_repository.dart';
