+++ lib/src/common/data/repositories/report_repository_impl.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:vocabualize/src/common/data/data_sources/remote_database_data_source.dart';
+import 'package:vocabualize/src/common/data/sources/remote_database_data_source.dart';
 import 'package:vocabualize/src/common/data/mappers/report_mappers.dart';
 import 'package:vocabualize/src/common/domain/entities/report.dart';
 import 'package:vocabualize/src/common/domain/repositories/report_repository.dart';
