+++ lib/src/common/data/repositories/vocabulary_repository_impl.dart
@@ -2,7 +2,7 @@ import 'dart:async';
 
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:log/log.dart';
-import 'package:vocabualize/src/common/data/data_sources/remote_database_data_source.dart';
+import 'package:vocabualize/src/common/data/sources/remote_database_data_source.dart';
 import 'package:vocabualize/src/common/data/mappers/vocabulary_mappers.dart';
 import 'package:vocabualize/src/common/data/models/rdb_event_type.dart';
 import 'package:vocabualize/src/common/domain/entities/filter_options.dart';
