+++ lib/src/common/data/sources/authentication_data_source.dart
@@ -5,7 +5,7 @@ import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:pocketbase/pocketbase.dart';
 import 'package:vocabualize/constants/common_imports.dart';
 import 'package:log/log.dart';
-import 'package:vocabualize/src/common/data/data_sources/remote_connection_client.dart';
+import 'package:vocabualize/src/common/data/sources/remote_connection_client.dart';
 import 'package:vocabualize/src/common/presentation/widgets/connection_checker.dart';
 
 final authenticationDataSourceProvider = Provider((ref) {
