+++ lib/src/features/details/data/repositories/speech_to_text_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:speech_to_text/speech_to_text.dart';
-import 'package:vocabualize/src/common/data/data_sources/speech_to_text_data_source.dart';
-import 'package:vocabualize/src/common/domain/repositories/speech_to_text_repository.dart';
+import 'package:vocabualize/src/features/record/data/sources/speech_to_text_data_source.dart';
+import 'package:vocabualize/src/features/details/domain/repositories/speech_to_text_repository.dart';
 
 final speechToTextRepositoryProvider = Provider((ref) {
   return SpeechToTextRepositoryImpl(
