+++ lib/src/features/details/data/sources/stock_image_data_source.dart
