+++ lib/src/features/details/domain/repositories/speech_to_text_repository.dart
