+++ lib/src/features/details/presentation/controllers/details_controller.dart
@@ -3,8 +3,8 @@ import 'package:url_launcher/url_launcher.dart';
 import 'package:vocabualize/constants/common_imports.dart';
 import 'package:vocabualize/src/common/domain/entities/vocabulary.dart';
 import 'package:vocabualize/src/common/domain/entities/vocabulary_image.dart';
-import 'package:vocabualize/src/common/domain/usecases/image/get_draft_image_use_case.dart';
-import 'package:vocabualize/src/common/domain/usecases/image/get_stock_images_use_case.dart';
+import 'package:vocabualize/src/features/details/domain/usecases/image/get_draft_image_use_case.dart';
+import 'package:vocabualize/src/features/details/domain/usecases/image/get_stock_images_use_case.dart';
 import 'package:vocabualize/src/common/domain/usecases/settings/get_are_images_enabled_use_case.dart';
 import 'package:vocabualize/src/common/domain/usecases/translator/translate_to_english_use_case.dart';
 import 'package:vocabualize/src/common/domain/usecases/vocabulary/delete_vocabulary_use_case.dart';
