+++ lib/src/features/record/presentation/widgets/mic_button.dart
@@ -3,7 +3,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:vocabualize/constants/common_imports.dart';
 import 'package:provider/provider.dart' as provider;
 import 'package:vocabualize/src/common/domain/entities/vocabulary.dart';
-import 'package:vocabualize/src/common/domain/usecases/language/record_speech_use_case.dart';
+import 'package:vocabualize/src/features/details/domain/usecases/speech_to_text/record_speech_use_case.dart';
 import 'package:vocabualize/src/common/domain/usecases/translator/translate_use_case.dart';
 import 'package:vocabualize/src/common/presentation/widgets/connection_checker.dart';
 import 'package:vocabualize/src/features/details/presentation/screens/details_screen.dart';
